reworking input system
this shit was cringe too understand lmao, we removed all the new input system stuff and took no notes of how we done it before so ive started reworking it in a separate scene. all the new stuff ive been doing is in the "Inputs" folder and "Testing" scene. Player can input on a new input system and spawn in a player now.